release-CLI update (#580)
* logs

* dont add tag if slug present

* debug prints

* temp

* logs for get_created_at

* temp

* temp

* get_AppKey fix

* subapp release fix

* removed logs

* after rebase big fix

* after rebase bug fix

+++ supervisely/cli/release/run.py
@@ -22,7 +22,8 @@ from supervisely.cli.release.release import (
     slug_is_valid,
     delete_tag,
     get_instance_version,
-    get_user
+    get_user,
+    get_created_at,
 )
 
 
@@ -302,7 +303,10 @@ def run(
         return False
 
     # get appKey
-    appKey = get_appKey(repo, sub_app_directory)
+    remote_name = repo.active_branch.tracking_branch().remote_name
+    remote = repo.remote(remote_name)
+    repo_url = remote.url
+    appKey = get_appKey(repo, sub_app_directory, repo_url)
 
     # check if app exist or not
     module_exists_label = "[yellow bold]updated[/]"
@@ -342,11 +346,12 @@ def run(
     console.print(
         f'\nApplication "{app_name}" will be {module_exists_label} at "{server_address}" Supervisely instance with release [blue]{release_version}[/] "{release_description}"'
     )
-    if repo.active_branch.name in ["main", "master"]:
-        remote_name = repo.active_branch.tracking_branch().name
-        console.print(
-            f'Git tag "sly-release-{release_version}" will be added and pushed to remote "{remote_name}"'
-        )
+    if not slug:
+        if repo.active_branch.name in ["main", "master"]:
+            remote_name = repo.active_branch.tracking_branch().name
+            console.print(
+                f'Git tag "sly-release-{release_version}" will be added and pushed to remote "{remote_name}"'
+            )
 
     # ask for confiramtion if needed
     if autoconfirm:
@@ -358,22 +363,27 @@ def run(
         return False
 
     # add tag and push
+    tag_name = None
     tag_created = False
-    if repo.active_branch.name in ["main", "master"]:
-        tag_name = f"sly-release-{release_version}"
-        tag = find_tag_in_repo(tag_name, repo)
-        if tag is None:
-            tag_created = True
-            repo.create_tag(tag_name)
-        try:
-            push_tag(tag_name, repo)
-        except subprocess.CalledProcessError:
-            if tag_created:
-                repo.delete_tag(tag)
-            console.print(
-                f"[red][Error][/] Git push unsuccessful. You need write permissions in repository to release the application"
-            )
-            return False
+    if not slug:
+        if repo.active_branch.name in ["main", "master"]:
+            tag_name = f"sly-release-{release_version}"
+            tag = find_tag_in_repo(tag_name, repo)
+            if tag is None:
+                tag_created = True
+                repo.create_tag(tag_name, message=release_description)
+            try:
+                push_tag(tag_name, repo)
+            except subprocess.CalledProcessError:
+                if tag_created:
+                    repo.delete_tag(tag)
+                console.print(
+                    f"[red][Error][/] Git push unsuccessful. You need write permissions in repository to release the application"
+                )
+                return False
+            
+    # get created_at
+    created_at = get_created_at(repo, tag_name)
 
     # release
     if release_token:
@@ -393,6 +403,7 @@ def run(
         slug,
         user_id,
         sub_app_directory if sub_app_directory != None else "",
+        created_at
     )
     if response.status_code != 200:
         error = f"[red][Error][/] Error releasing the application. Please contact Supervisely team. Status Code: {response.status_code}"
